third_party/WebKit/LayoutTests/svg/css/getComputedStyle-svg-text-width-height.html

<!DOCTYPE html>
<title>Width/height properties in getComputedStyle of svg text</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg width="100" height="100">
  <text id="text1">Text1</text>
  <text id="text2" style="width: 20px; height: 30%">Text2</text>
</svg>
<script>
test(function() {
  assert_equals(getComputedStyle(text1).width, 'auto');
  assert_equals(getComputedStyle(text1).height, 'auto');
  assert_equals(getComputedStyle(text2).width, 'auto');
  assert_equals(getComputedStyle(text2).height, 'auto');
});
</script>
